+++ templates/element/flash/flash.php
@@ -14,9 +14,8 @@ declare(strict_types=1);
  */
 ?>
 
-<div class="alert <?= $params['class'] ?? null ?> alert-dismissible border-0 fade show" role="alert">
-    <button type="button" class="close" data-dismiss="alert" aria-label="<?= __d('me_tools', 'Close') ?>">
-        <span aria-hidden="true">&times;</span>
-    </button>
+<div class="alert alert-dismissible <?= $params['class'] ?? null ?> p-3 fade show" role="alert">
+    <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="<?= __d('me_tools', 'Close') ?>"></button>
+
     <?= $message ?>
 </div>