+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 2.21.1
 * fixed a little bug for `HtmlHelper::link()`;
 * little fixes for `FormHelper`, to make the output more consistent with what
-    Bootstrap requires.
+    Bootstrap requires;
+* fixed the `flash` template element.
 
 ### 2.21.0
 * numerous code adjustments for improvement and adaptation to PHP 7.4 new features;